// src/features/tables/types/table.d.ts
export interface TableResponse {
    tableId: number;
    name: string;
    isOccupied: boolean;
    seats: number;
    location: string;
    restaurantId: number;
}

export interface CreateTableDto {
    name: string;
    seats: number;
    location: string;
}

export interface UpdateTableDto extends Partial<CreateTableDto> {
    tableId: number;
}

export interface DeleteTableDto {
    tableId: number
}